refactor(audio): rename and enhance audio speed change functionality with improved tests
