+++ src/pages/tools/audio/change-speed/change-speed.service.test.ts
@@ -16,11 +16,11 @@ vi.mock('@ffmpeg/util', () => ({
   fetchFile: vi.fn().mockResolvedValue(new Uint8Array([1, 2, 3, 4, 5]))
 }));
 
-import { main } from './service';
+import { changeAudioSpeed } from './service';
 import { InitialValuesType } from './types';
 
-describe('changeSpeed (main)', () => {
-  it('should return the input file unchanged (mock implementation)', () => {
+describe('changeAudioSpeed', () => {
+  it('should return a new File with the correct name and type', async () => {
     const mockAudioData = new Uint8Array([0, 1, 2, 3, 4, 5]);
     const mockFile = new File([mockAudioData], 'test.mp3', {
       type: 'audio/mp3'
@@ -29,7 +29,18 @@ describe('changeSpeed (main)', () => {
       newSpeed: 2,
       outputFormat: 'mp3'
     };
-    const result = main(mockFile, options);
-    expect(result).toBe(mockFile);
+    const result = await changeAudioSpeed(mockFile, options);
+    expect(result).toBeInstanceOf(File);
+    expect(result?.name).toBe('test-2x.mp3');
+    expect(result?.type).toBe('audio/mp3');
+  });
+
+  it('should return null if input is null', async () => {
+    const options: InitialValuesType = {
+      newSpeed: 2,
+      outputFormat: 'mp3'
+    };
+    const result = await changeAudioSpeed(null, options);
+    expect(result).toBeNull();
   });
 });
