+++ src/pages/tools/audio/merge-audio/service.ts
@@ -22,65 +22,86 @@ export async function mergeAudioFiles(
   const { outputFormat } = options;
   const outputName = `output.${outputFormat}`;
 
-  // Write all input files to FFmpeg
-  const inputNames: string[] = [];
+  // 1. Convert all inputs to WAV
+  const tempWavNames: string[] = [];
   for (let i = 0; i < inputs.length; i++) {
-    const inputName = `input${i}.mp3`;
+    const inputName = `input${i}`;
+    const tempWavName = `temp${i}.wav`;
     await ffmpeg.writeFile(inputName, await fetchFile(inputs[i]));
-    inputNames.push(inputName);
-  }
-
-  // Create a file list for concatenation
-  const fileListName = 'filelist.txt';
-  const fileListContent = inputNames.map((name) => `file '${name}'`).join('\n');
-  await ffmpeg.writeFile(fileListName, fileListContent);
-
-  // Build FFmpeg arguments for merging
-  const args: string[] = ['-f', 'concat', '-safe', '0', '-i', fileListName];
-
-  // Add format-specific arguments
-  if (outputFormat === 'mp3') {
-    args.push(
-      '-ar',
-      '44100',
-      '-ac',
-      '2',
-      '-b:a',
-      '192k',
-      '-f',
-      'mp3',
-      outputName
-    );
-  } else if (outputFormat === 'aac') {
-    args.push('-c:a', 'aac', '-b:a', '192k', '-f', 'adts', outputName);
-  } else if (outputFormat === 'wav') {
-    args.push(
+    await ffmpeg.exec([
+      '-i',
+      inputName,
       '-acodec',
       'pcm_s16le',
       '-ar',
       '44100',
       '-ac',
       '2',
-      '-f',
-      'wav',
-      outputName
-    );
-  }
-
-  await ffmpeg.exec(args);
-
-  const mergedAudio = await ffmpeg.readFile(outputName);
-
-  let mimeType = 'audio/mp3';
-  if (outputFormat === 'aac') mimeType = 'audio/aac';
-  if (outputFormat === 'wav') mimeType = 'audio/wav';
-
-  // Clean up files
-  for (const inputName of inputNames) {
+      tempWavName
+    ]);
+    tempWavNames.push(tempWavName);
     await ffmpeg.deleteFile(inputName);
   }
+
+  // 2. Create file list for concat
+  const fileListName = 'filelist.txt';
+  const fileListContent = tempWavNames
+    .map((name) => `file '${name}'`)
+    .join('\n');
+  await ffmpeg.writeFile(fileListName, fileListContent);
+
+  // 3. Concatenate WAV files
+  const concatWav = 'concat.wav';
+  await ffmpeg.exec([
+    '-f',
+    'concat',
+    '-safe',
+    '0',
+    '-i',
+    fileListName,
+    '-c',
+    'copy',
+    concatWav
+  ]);
+
+  // 4. Convert concatenated WAV to requested output format
+  let finalOutput = concatWav;
+  if (outputFormat !== 'wav') {
+    const args = ['-i', concatWav];
+    if (outputFormat === 'mp3') {
+      args.push(
+        '-ar',
+        '44100',
+        '-ac',
+        '2',
+        '-b:a',
+        '192k',
+        '-f',
+        'mp3',
+        outputName
+      );
+    } else if (outputFormat === 'aac') {
+      args.push('-c:a', 'aac', '-b:a', '192k', '-f', 'adts', outputName);
+    }
+    await ffmpeg.exec(args);
+    finalOutput = outputName;
+  }
+
+  const mergedAudio = await ffmpeg.readFile(finalOutput);
+
+  let mimeType = 'audio/wav';
+  if (outputFormat === 'mp3') mimeType = 'audio/mp3';
+  if (outputFormat === 'aac') mimeType = 'audio/aac';
+
+  // Clean up files
+  for (const tempWavName of tempWavNames) {
+    await ffmpeg.deleteFile(tempWavName);
+  }
   await ffmpeg.deleteFile(fileListName);
-  await ffmpeg.deleteFile(outputName);
+  await ffmpeg.deleteFile(concatWav);
+  if (outputFormat !== 'wav') {
+    await ffmpeg.deleteFile(outputName);
+  }
 
   return new File(
     [
